enable message size larger than 4GB for all_to_all_v/all_gather_v (#228)

+++ tutel/examples/nccl_all_to_all_v.py
@@ -12,10 +12,10 @@ assert parallel_env.global_size == 2, "This test case is set for World Size == 2
 
 if parallel_env.global_rank == 0:
   input = torch.tensor([10, 10, 10, 10, 10], device=local_device)
-  send_counts = torch.tensor([1, 4], dtype=torch.int32, device=local_device)
+  send_counts = torch.tensor([1, 4], dtype=torch.int64, device=local_device)
 else:
   input = torch.tensor([20, 20, 20], device=local_device)
-  send_counts = torch.tensor([2, 1], dtype=torch.int32, device=local_device)
+  send_counts = torch.tensor([2, 1], dtype=torch.int64, device=local_device)
 
 print(f'Device-{parallel_env.global_rank} sends: {[input,]}')
